+++ src/editor/lib/MonacoEditor.js
@@ -1,5 +1,4 @@
 import { setupMonaco, getAppropriateMonacoTheme } from './monacoEditorHelper';
-import * as util from '../../global/util';
 
 let configuredPromise;
 
@@ -30,7 +29,7 @@ export class MonacoEditor {
             // on apex 20.2, on the plug-in page, there would be a conflict with the apex monaco
             // for now, we won't be able to use the fancy highlighting for js, as it would break the apex plsql code editor
             // so we don't initialize it, which means we'll be using the apex monaco instance
-            if (!configuredPromise && util.apexVersion == '20.2' && util.pageId == 4410) {
+            if (!configuredPromise && FOS.util.apexVersion == '20.2' && FOS.util.pageId == 4410) {
                 // dummy promise
                 configuredPromise = new Promise(function(resolve){
                     resolve();