+++ src/editor/lib/server.js
@@ -1,10 +1,8 @@
-import * as util from '../../global/util.js';
-
 // files should be passed on as: [{directory, fileName, mimeType, content}]
 // returns a promise that resolves to an object of {ok: boolean, error: string}
 const uploadPluginFiles = async function(uploadPageUrl, files){
 
-    let pageId = util.pageId;
+    let pageId = FOS.util.pageId;
 
     switch(pageId){
         case 40: