+++ src/global/script.js
@@ -1,0 +1,6 @@
+import * as util from './util';
+
+window.FOS = {};
+FOS.util = util;
+
+FOS.version = '20.2.0';